+++ OpenReferralApi/Services/Interfaces/IDashboardService.cs
@@ -7,8 +7,8 @@ namespace OpenReferralApi.Services.Interfaces;
 
 public interface IDashboardService
 {
-    public Task<Result<DashboardOutput>> GetServices();
-    public Task<Result<DashboardServiceDetails>> GetServiceById(string id);
+    public Task<Result<DashboardResponse>> GetServices();
+    public Task<Result<ServiceDetailsResponse>> GetServiceById(string id);
     public Task<Result<List<DashboardValidationResponse>>> ValidateDashboardServices();
     public Task<Result<SubmissionResponse>> SubmitService(DashboardSubmissionRequest submission);
 }
